Move cursors to ui.h.
