+++ src/ui/button.c
@@ -3,8 +3,9 @@
 #include "draw.h"
 #include "tooltip.h"
 
-#include "../main.h"
 #include "../macros.h"
+#include "../main.h"
+#include "../ui.h"
 
 static void calculate_pos_and_width(BUTTON *b, int *x, int *w) {
     int old_w = *w;
