+++ src/windows/main.c
@@ -10,9 +10,10 @@
 #include "../main.h"
 #include "../main_native.h"
 #include "../settings.h"
-#include "../theme.h"
 #include "../text.h"
+#include "../theme.h"
 #include "../tox.h"
+#include "../ui.h"
 #include "../utox.h"
 
 #include "../av/utox_av.h"
