+++ src/xlib/main.c
@@ -7,6 +7,7 @@
 #include "../friend.h"
 #include "../logging_native.h"
 #include "../macros.h"
+#include "../main_native.h"
 #include "../settings.h"
 #include "../theme.h"
 #include "../tox.h"
